Fix parallel markduplicate function
The way the glob was constructed caused multiple files from the same
sample, but from different libraries, to be grouped together for merging.
The reason was that the glob extended everything after the library
identifier, without heeding the separator _. This caused samples like
A_S1_L001 / A_S1_L002 (which should be grouped) to be grouped with
A_S11_L001 (which is a separate library that should be merged along its
own lanes separately).