Allow setting non-existing directory as output target
The output directory provided on the CLI was converted to its
full path via the `realpath` function. However, realpath throws
an error when parent-level components of the path do not exist.
In this case, if an output directory is chosen such as
"new_dir/output", then the `output_dir` variable inside the script
could not be set. By adding the `-m`/`--canonicalize-missing` flag
to realpath, this error is avoided.