Refactor parallel markduplicates step into separate function
Uses a separate function for calling markduplicates  in order
to make it easier to use it in combination with gnu parallel.
E.g., now it should be easier to provide logging messages
for each bam file, processed files can be skipped, and it is easier
to reference the current bam file without complex behaviour
of quotes/escapes.