Add joint variant calling and annotation for multiple directories
Adds scripts to perform the joint variant calling steps
(genomicsdbimport and GenotypeGVCFs) on multiple directories
with gvcf files, by providing a list of species and then using
find to retrieve all of them.

The same is done for the joint annotation script, so that it can
work in absence of a samplesheet.